fixed the test
